+++ var/httpd/htdocs/js/Core.UI.InputFields.js
@@ -399,7 +399,7 @@ Core.UI.InputFields = (function (TargetNS) {
 
         // Check for empty values (allow field clearing).
         $SelectObj.find('option').each(function (Index, Option) {
-            if ($(Option).attr('value') === '' || $(Option).attr('value') === '||-') {
+            if ($(Option).attr('value') === ' ' ||  $(Option).attr('value') === '' || $(Option).attr('value') === '||-') {
                 PossibleNone = true;
                 return true;
             }
@@ -1464,7 +1464,7 @@ Core.UI.InputFields = (function (TargetNS) {
                     }
 
                     $SelectObj.find('option').each(function (Index, Option) {
-                        if ($(Option).attr('value') === '' || $(Option).attr('value') === '||-') {
+                        if ($(Option).attr('value') === ' ' ||  $(Option).attr('value') === '' || $(Option).attr('value') === '||-') {
                             PossibleNone = true;
                             return true;
                         }